+++ web/src/beta/features/Editor/tabs/story/PageIndicator/index.tsx
